Add callback from parent to child
You can't modify the props that are passed from parent to child in the child component
so because its like copying not like a pointer. so how can you modify the parent state in the child state? by
passing callback function from parent to child and in that way you can chnage the state of a var in the parent componet thorugh that callback function